Move iterator types to separate module
